20230209 优化隐藏网络连接,如搜索不到或连接不上隐藏网络，请断开现有连接或关闭重新打开wifi后再次连接隐藏网络

+++ AirPortOpenBSD/AirPort_OpenBSD.hpp
@@ -343,9 +343,9 @@ public:
     unsigned long currentPMPowerLevel = 0;
     
     static IOReturn setPowerStateAction(OSObject *owner, void *arg1, void *arg2, void *arg3, void *arg4);
-    IOReturn changePowerState(OSObject *object, int powerState);
-    unsigned long powerState;
-    u_int32_t    powersave_level;
+    IOReturn changePowerState(OSObject *object, u_int32_t powerState);
+    u_int32_t   powerState;
+    u_int32_t   powersave_level;
     struct apple80211_antenna_data tx_antenna;
     struct apple80211_antenna_data ad;
